refactorting was necessary since I realised I need the same data at more pages, so I lifted it to _app.js and now handing down the data
