+++ components/BarList/index.js
@@ -1,0 +1,16 @@
+import { bars } from "../../lib/mock-data/bars";
+
+import { teams } from "../../lib/mock-data/teams";
+import { barsInMatches } from "../../lib/mock-data/barsInMatches";
+
+
+export default function BarList({matches}){
+
+    console.log("test:",matches)
+    
+    return(
+        <ul>
+            <li>test</li>
+        </ul>
+    )
+}
