+++ pages/bars/index.js
@@ -1,24 +1,15 @@
-import { matches } from "../../lib/mock-data/matches";
-import { teams } from "../../lib/mock-data/teams";
 import AppHeader from "../../components/AppHeader/AppHeader";
 import AppFooter from "../../components/AppFooter/AppFooter";
+import BarList from "../../components/BarList";
 
 
-export default function Bars(){
-    const matchesWithTeamNames = matches.map((match) => (
-        {...match, 
-            homeTeam: {name: teams.find((team) => team.id === match.homeTeamId).name, 
-            logoColor: teams.find((team) => team.id === match.homeTeamId).logoColor}, 
-            awayTeam:{name: teams.find((team) => team.id === match.awayTeamId).name, 
-            logoColor: teams.find((team) => team.id === match.awayTeamId).logoColor}}
-    ))
-
-
+export default function Bars({matches}){
 
     return(
         <>
         <AppHeader/>
         <h1>Bars Overview</h1>
+        <BarList matches={matches}/>
         <AppFooter/>
         </>
 )
