+++ pages/index.js
@@ -2,12 +2,12 @@ import MatchList from "../components/MatchList";
 import AppHeader from "../components/AppHeader/AppHeader";
 import AppFooter from "../components/AppFooter/AppFooter";
 
-export default function Home() {
+export default function Home({matches}) {
   return (
     <>
     <AppHeader/>
       <main>
-        <MatchList/>
+        <MatchList matches={matches}/>
       </main>
     <AppFooter/>
     </>
